[장바구니 미션 Step 2] 첵스(최혜령) 미션 제출합니다. (#206)
* chore: 배포 설정 변경

* refactor: 상품목록 fetch하도록 수정

* fix: exhaustive-deps-warning 해결

* chore: msw 적용

* chore: 스토리북 불필요한 설정 코드 삭제

* fix: storybook mockServiceWorker 경로 오류 해결

* rename: 커스텀 훅 함수명 변경

* rename: Product, ProductList를 Product폴더로 이동

* feat: Price 컴포넌트 생성 및 스토리북 레이아웃 설정

* rename: productsInCart -> cartList로 변수명 변경

* feat: CartItem 컴포넌트 생성

* refactor: 장바구니 목록에 담긴 품목 수량 selector로 생성 후 컴포넌트 분리

* design: Button 컴포넌트의 text 디자인타입 추가

* chore: 스토리북 preview의 parameters에 msw: handlers 추가

* refactor: 스토리북 Stepper 테스트 실패 코드 수정

* feat: CartList 컴포넌트 생성

* style: css 수정 및 불필요한 코드 삭제

* feat: TotalPayment 컴포넌트 및 장바구니 페이지 생성

* style: 장바구니가 비었을 때 css 수정

* refactor: useEffect 대신 atomFamily를 이용하여 초기화

useEffect 대신 atomFamily에 장바구니목록의 id들을 전달 후 checkedItemIds 상태를 초기화하도록 수정

* style: 이미지 비율 맞춰서 정사각형으로 자르기

* style: 태블릿, 모바일 환경 대응

* docs: README 업데이트

* feat: msw get, post, patch, delete 적용

* style: props 구조 분해 할당 적용 및 기타 코드 수정

* feat: 로딩 시 Suspense 적용 및 컴포넌트 마운트 후 상태 업데이트 하도록 수정

* refactor: 웹 접근성 개선

* feat: 404페이지 추가

* docs: README 2단계 실행화면 추가

* refactor: 불필요한 상태(isAllChecked) 제거

* refactor: Checkbox를 CartItem 내에서 관리하도록 수정

* refactor: 선택된 아이템이 없는 경우 주문하기 및 선택삭제 버튼 비활성화

* refactor: 장바구니 목록 fetch 로직 selector 밖으로 분리

* refactor: TotalPayment 스토리북 스토리 추가

* design: 버튼 컴포넌트 disabled 시 보여주는 폰트 색 변경

+++ src/hooks/useCart.ts
@@ -1,49 +1,50 @@
+import { useEffect } from 'react';
 import { useRecoilState } from 'recoil';
-import { productsInCartState } from '../recoil/atoms';
+import { cartListState } from '../recoil/atoms';
+import { CartItemInfo, ProductInfo } from '../types';
+import { CART_BASE_URL } from '../constants';
+import { useSetFetchedData } from './useSetFetchedData';
 
-export const useCart = (productId: number) => {
-  const [productsInCart, setProductsInCart] = useRecoilState(productsInCartState);
+export const useCart = (productInfo?: ProductInfo) => {
+  const [cartList, setCartList] = useRecoilState(cartListState);
+  const { api } = useSetFetchedData<CartItemInfo[]>(CART_BASE_URL, setCartList);
 
-  const getProductInCart = () => {
-    return productsInCart.find((product) => product.id === productId);
+  const setDefaultCartList = () => {
+    api.get(CART_BASE_URL);
+  };
+
+  const getCartItem = () => {
+    return cartList.find((cartItem) => cartItem.id === productInfo?.id);
   };
 
   const addToCart = () => {
-    setProductsInCart((prev) => [
-      ...prev,
-      {
-        id: productId,
-        quantity: 1,
-      },
-    ]);
-  };
-
-  const deleteFromCart = () => {
-    const curIndex = productsInCart.findIndex((product) => product.id === productId);
-    const newProductsInCart = structuredClone(productsInCart);
-
-    newProductsInCart.splice(curIndex, 1);
-    setProductsInCart(newProductsInCart);
+    if (!productInfo) return;
+    api.post(CART_BASE_URL, { productId: productInfo.id }, CART_BASE_URL);
   };
 
   const updateProductQuantity = (quantity: number) => {
+    if (!productInfo) return;
     if (quantity <= 0) {
       deleteFromCart();
       return;
     }
 
-    setProductsInCart((prev) => {
-      return prev.map((productInCart) => {
-        if (productInCart.id === productId) return { ...productInCart, quantity: quantity };
-
-        return productInCart;
-      });
-    });
+    api.patch(`${CART_BASE_URL}/${productInfo.id}`, { quantity: quantity }, CART_BASE_URL);
   };
 
+  const deleteFromCart = (productId?: number) => {
+    const curProductId = productId ? productId : productInfo?.id;
+    api.delete(`${CART_BASE_URL}/${curProductId}`, CART_BASE_URL);
+  };
+
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  useEffect(() => setDefaultCartList(), []);
+
   return {
-    productsInCart,
-    getProductInCart,
+    cartList,
+    setCartList,
+    setDefaultCartList,
+    getCartItem,
     addToCart,
     deleteFromCart,
     updateProductQuantity,
